+++ tool_packages/unique_web_search/src/unique_web_search/services/search_engine/base.py
@@ -2,12 +2,9 @@ from abc import ABC, abstractmethod
 from enum import StrEnum
 from typing import Generic, TypeVar
 
-from pydantic import BaseModel, Field
+from pydantic import BaseModel, ConfigDict, Field
 from unique_toolkit.agentic.tools.config import get_configuration_dict
 
-from unique_web_search.services.helpers import (
-    clean_model_title_generator,
-)
 from unique_web_search.services.search_engine.schema import (
     WebSearchResult,
 )
@@ -25,15 +22,35 @@ class SearchEngineType(StrEnum):
     CUSTOM_API = "CustomAPI"
 
 
+_SearchEngineExposedName = {
+    SearchEngineType.GOOGLE: "Google Search Engine",
+    SearchEngineType.JINA: "Jina Search",
+    SearchEngineType.FIRECRAWL: "Firecrawl Search",
+    SearchEngineType.TAVILY: "Tavily Search",
+    SearchEngineType.BRAVE: "Brave Search Engine",
+    SearchEngineType.BING: "Grounding with Bing",
+    SearchEngineType.DUCKDUCKGO: "DuckDuckGo Search Engine",
+    SearchEngineType.VERTEXAI: "VertexAI Search Engine",
+    SearchEngineType.CUSTOM_API: "Customized API Search Engine",
+}
+
+
+def get_search_engine_model_config(
+    search_engine_name: SearchEngineType,
+) -> ConfigDict:
+    return get_configuration_dict(
+        title=_SearchEngineExposedName.get(
+            search_engine_name, "Undefined Search Engine"
+        )
+    )
+
+
 T = TypeVar("T", bound=SearchEngineType)
 
 
 class BaseSearchEngineConfig(BaseModel, Generic[T]):
-    model_config = get_configuration_dict(
-        model_title_generator=clean_model_title_generator
-    )
+    model_config = get_configuration_dict()
     search_engine_name: T
-
     fetch_size: int = Field(
         default=5,
         description="Number of search results to fetch",